requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
argon2-cffi==23.1.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
pytest-cov==4.1.0
